--- components/standing/PlayerinfoPicture.tsx
@@ -1,9 +1,0 @@
-import styles from "./PlayerInfo.module.css";
-
-export type User = {
-  picture: string;
-};
-
-export default function UserPicture({ picture }: User) {
-  return <img className={styles.playerInfoPicture} src={picture} />;
-}
